requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database & ORM
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4

# Pydantic
pydantic==2.7.1
pydantic-settings==2.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
cryptography==42.0.7

# Messaging & Events
aio-pika==9.4.1
kafka-python==2.0.2

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9

# AI, ML & Data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
google-generativeai==0.7.1
sentence-transformers==2.7.0

# Vector DB
pymilvus==2.4.1

# Analytics DB
clickhouse-driver==0.2.7

# Object Storage
boto3==1.34.102

# Monitoring
prometheus-client==0.20.0